// src/feature/auth/model/ProvinceEnum.ts
export enum ProvinceEnum {
    nad = "Nanggroe Aceh Darussalam",
    northSumatera = "Sumatera Utara",
    southSumatera = "Sumatera Selatan",
    westSumatera = "Sumatera Barat",
    bengkulu = "Bengkulu",
    riau = "Riau",
    riauKepualauan = "Kepulauan Riau",
    jambi = "Jambi",
    lampung = "Lampung",
    bangkaBelitung = "Bangka Belitung",
    westKalimantan = "Kalimantan Barat",
    eastKalimantan = "Kalimantan Timur",
    southKalimantan = "Kalimantan Selatan",
    centralKalimantan = "Kalimantan Tengah",
    northKalimantan = "Kalimantan Utara",
    banten = "Banten",
    jakarta = "DKI Jakarta",
    westJava = "Jawa Barat",
    centralJava = "Jawa Tengah",
    yogyakarta = "Daerah Istimewa Yogyakarta",
    eastJava = "Jawa Timur",
    bali = "Bali",
    ntt = "Nusa Tenggara Timur",
    ntb = "Nusa Tenggara Barat",
    gorontalo = "Gorontalo",
    westSulawesi = "Sulawesi Barat",
    centralSulawesi = "Sulawesi Tengah",
    northSulawesi = "Sulawesi Utara",
    southeastSulawesi = "Sulawesi Tenggara",
    southSulawesi = "Sulawesi Selatan",
    northMaluku = "Maluku Utara",
    maluku = "Maluku",
    westPapua = "Papua Barat",
    papua = "Papua",
    centralPapua = "Papua Tengah",
    papuaPegunungan = "Papua Pegunungan",
    southPapua = "Papua Selatan",
    southwestPapua = "Papua Barat Daya",
}